+++ sphinx_astropy/conf/v1.py
@@ -136,7 +136,8 @@ extensions = [
     'sphinx_automodapi.automodapi',
     'sphinx_automodapi.smart_resolver',
     'sphinx_astropy.ext.doctest',
-    'sphinx_astropy.ext.changelog_links']
+    'sphinx_astropy.ext.changelog_links',
+    'sphinx_astropy.ext.intersphinx_toggle']
 
 
 if not on_rtd and LooseVersion(sphinx.__version__) < LooseVersion('1.4'):